feat(general): Cleanup section code (#40)
This refactors and cleans up the code for the section logic (This will
affect our other open prs but I am happy to modify them to fit this new
style). We now require styles to be built through the section element
but this means we can localize our style logic which keeps things more
consistent, additionally I reduced the complexity of the css for our
other sections which helps with keeping our code clean and shipping
small css bundles.

This code also changes `NewsSection` to support items with an `href` as
we will not have the blog to link to when we first release and we might
just want to show news elements without more detail.

Additionally this code makes a new `utils.ts` file which we can use to
store global utilties such as `formatDate` allowing for consistent use
and less code duplication.